feat: contracts and business model types are defined

+++ src/internal/logic/rewards/domain/types.go
@@ -1,7 +1,7 @@
 package domain
 
 type RewardsAccumulateResponse struct {
-	RewardsEarned   int
+	RewardsEarned   float64
 	RewardsType     string
 	CampaignApplied bool
 }
